Create preview icon for Icons entry point (2/3)
Create the preview icon by retrieving the adaptive icon from the app to
be previewed, then wrapping it in a ShapeIconDrawable, which handles
drawing it in the given icon shape, and displaying the monochrome icon
with correct theme colors if themed. Make sure the colors are updated
also when the system color changes.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 402161932
Test: manually verified
Test: AppIconPickerViewModelTest
Change-Id: I2f7138de1f2b540df943f1306a6301b5d8922ce0