Hide clock click delegate view for A11y when not interactable (2/2)
Clock face click delegate view is not interactable outside of clock
picker, and when the selected clock is not interactable, therefore it
should be hidden.

Also make sure view content description is set correctly on the view,
for the case that the default clock is selected and it should be shown.

Flag: com.android.systemui.shared.new_customization_picker_ui
Bug: 418202194
Test: manually verified with Talkback, switch access
Change-Id: I4c0bfd9b1321ecbe82b695a436dc04f743bdca5e