Fix clock color container height
In http://ag/33867509 we introduced a bug where when there are two rows of color options, the slider will be pushed down and cut off by the container. This was due to the invisible placeholder that always fills the height of the color container with a fixed height that fits one row of color options.

This CL fixes it by changing the placeholder to the height of two rows of color options on desktop.

Screenshot (Desktop UI on): http://go/scrcast/NTY1MjYyODc3OTMwMjkxMnwzYzg5YzdlZS05Nw
Screenshot (Desktop UI off): http://go/scrcast/NTQyMTkxMzE2ODkzNjk2MHw5ZjA3NjMyYi1kNw

Bug: 415902512
Test: Manually on DUT
Flag: EXEMPT Bug fixes for desktop
Change-Id: I2debf653036df05c50f81197f15da569e54e4ba3