Refactor monet.Style to ThemeStyle
Deprecates `monet.Style` in favor of the new `ThemeStyle` class in preparation for the upcoming `ThemeManagerService`.

This change is purely a refactoring, moving the existing `monet.Style` class to a new location and name with no functional changes. All existing usages in SystemUI have been updated to the new class.

Bug: 441912053
Test: Existing tests updated.
Flag: EXEMPT mechanical refactoring
Change-Id: Icece6013442d184b41ed83d1d6b0393eedf4e027